+++ 1-server-side/msal-webapi-ciam-sample/msal-web-api/src/main/java/com/microsoft/azure/msalapisample/AuthException.java
@@ -1,15 +1,11 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.microsoft.azure.msalwebsample;
+package com.microsoft.azure.msalapisample;
 
 public class AuthException extends RuntimeException {
 
     public AuthException(String message, Throwable cause){
         super(message, cause);
-    }
-
-    public AuthException(String message){
-        super(message);
     }
 }
