+++ 1-server-side/msal-webapi-ciam-sample/msal-web-api/src/main/java/com/microsoft/azure/msalapisample/CachingConfig.java
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.microsoft.azure.msalobosample;
+package com.microsoft.azure.msalapisample;
 
 import com.google.common.cache.CacheBuilder;
 import org.springframework.cache.Cache;
