+++ 1-server-side/msal-webapi-ciam-sample/msal-web-app/src/main/java/com/microsoft/azure/msalappsample/AuthException.java
@@ -1,11 +1,15 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.microsoft.azure.msalobosample;
+package com.microsoft.azure.msalappsample;
 
 public class AuthException extends RuntimeException {
 
     public AuthException(String message, Throwable cause){
         super(message, cause);
+    }
+
+    public AuthException(String message){
+        super(message);
     }
 }
