+++ 1-server-side/msal-webapi-ciam-sample/msal-web-app/src/main/java/com/microsoft/azure/msalappsample/AuthHelper.java
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.microsoft.azure.msalwebsample;
+package com.microsoft.azure.msalappsample;
 
 import java.net.URI;
 import java.util.Arrays;
