+++ 1-server-side/msal-webapi-ciam-sample/msal-web-app/src/main/java/com/microsoft/azure/msalappsample/AuthPageController.java
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.microsoft.azure.msalwebsample;
+package com.microsoft.azure.msalappsample;
 
 import com.microsoft.aad.msal4j.IAuthenticationResult;
 import com.nimbusds.jwt.JWTParser;
@@ -21,7 +21,7 @@ import java.io.IOException;
 import java.net.URLEncoder;
 import java.text.ParseException;
 
-import static com.microsoft.azure.msalwebsample.AuthHelper.getAuthSessionObject;
+import static com.microsoft.azure.msalappsample.AuthHelper.getAuthSessionObject;
 
 @Controller
 public class AuthPageController {
